+++ api/src/types.ts
@@ -15,20 +15,23 @@ export type ProcessingItemType = {
 };
 
 export type TiddlConfig = {
-  token: string;
-  refresh_token: string;
-  token_expires_at: number;
-  settings: {
-    download_path: string;
-    track_quality: "master" | "high";
-    track_template: string;
-    album_template: string;
-    playlist_template: string;
-    file_extension: string;
-  };
-  user: {
+  auth: {
+    token: string;
+    refresh_token: string;
+    token_expires_at: number;
     user_id: string;
     country_code: string;
+  };
+  template: {
+    track: string;
+    video: string;
+    album: string;
+    playlist: string;
+  };
+  download: {
+    quality: "master" | "high";
+    path: string;
+    threads: number;
   };
 };