+++ e2e/tests/utils/search.ts
@@ -1,6 +1,10 @@
 import { expect, Page } from "@playwright/test";
 
-import { waitForImgLoaded, waitForLoader } from "./helpers";
+import {
+  emptyProcessingList,
+  waitForImgLoaded,
+  waitForLoader,
+} from "./helpers";
 import { mockConfigAPI, mockRelease } from "./mock";
 
 export async function runSearch(keyword: string, page: Page) {
@@ -22,6 +26,7 @@ export async function runSearch(keyword: string, page: Page) {
   await page.getByTestId("search-input").locator("input").press("Enter");
   await waitForLoader(page);
   await waitForImgLoaded(page);
+  await emptyProcessingList(page);
 }
 
 export async function countItems(